refactor trust basis generation and classifiers
